Rajneesh/feature/navigation architecture 010524

+++ app/src/main/java/com/example/ticketpartner/feature_scan_module/feature_qr_scan/data/datasource/QrScanDataSourceImpl.kt
@@ -35,7 +35,7 @@ class QrScanDataSourceImpl @Inject constructor(private val restApiService: RestA
     }
 
     override suspend fun getScanCheckedInResponse(
-        checkedOrderIdList: ArrayList<Int>,
+        checkedOrderIdList: ArrayList<Long>,
         orderId: String
     ): ScanCheckedInResponse {
         return restApiService.getQrScanCheckedInData(ScanCheckedInRequest(checkedOrderIdList,orderId))
